Fix: 컴포넌트 폴더관리 및 View컴포넌트 생성
