+++ components/store_select_and_detail/Menu.tsx
