+++ components/store_select_and_detail/StarRate.tsx
