+++ components/store_select_and_detail/StoreLists.tsx
