+++ pages/store/index.tsx
@@ -1,0 +1,5 @@
+import StoreListPageView from '@/components/views/store_select_and_detail_view/StoreListPageView';
+
+export default function StoreListPage() {
+  return <StoreListPageView />;
+}
